Dev: Refactor the application to use Vue.js
All of the UI logic is moved to use Vue.JS data binding,
removing all of the code interacting directly with the DOM.
The application is much shorter as a result.

Also, show the UI only when the data has been fetched, which
is stricly speaking another feature but was so easy to do in Vue
that I couldn't resist.